Move component path
