+++ src/components/LatestPosts.js
